backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dateutil
python-multipart
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
openpyxl
pandas
psycopg2-binary
python-dotenv
httpx
pytest
pytest-asyncio
pytest-cov
redis
slowapi
stripe
mercadopago
